+++ resources/views/back/pages/auth/reset.blade.php
@@ -5,7 +5,7 @@
 <div class="page page-center">
       <div class="container container-tight py-4">
         <div class="text-center mb-4">
-          <a href="." class="navbar-brand navbar-brand-autodark"><img src="./back/static/logo.svg" alt="" height="36"></a>
+          <a href="." class="navbar-brand navbar-brand-autodark"><img src="{{ \App\Models\Setting::find(1)->blog_logo }}" alt="" height="36"></a>
         </div>
 
         @livewire('author-reset-form')