+++ resources/views/back/pages/settings.blade.php
@@ -33,6 +33,22 @@
             </div>
             <div class="tab-pane fade" id="tabs-profile-8" role="tabpanel">
 
+                <div class="row">
+                    <div class="col-md-6">
+                        <h3>Set blog logo</h3>
+                        <div class="mb-2" style="max-width: 200px">
+                            <img src="" alt="" class="img-thumbnail" id="logo-img-preview" data-ijabo-default-img="{{ \App\Models\Setting::find(1)->blog_logo }}" >
+                        </div>
+                        <form action="{{ route('author.change-blog-logo')}}" method="POST" id="changeBlogLogoForm" >
+                            @csrf
+                            <div class="mb-2" >
+                                <input type="file" name="blog_logo" class="form-control">
+                            </div>
+                            <button class="btn btn-primary">Change logo</button>
+                        </form>
+                    </div>
+                </div>
+
             </div>
             <div class="tab-pane fade" id="tabs-activity-8" role="tabpanel">
 
@@ -43,4 +59,48 @@
 
 @endsection
 
+@push('scripts')
+  <script>
+    $('input[name="blog_logo"]').ijaboViewer({
+        preview     : '#logo-img-preview',
+        imageShape :  'rectangular',
+        allowedExtensions : ['jpg', 'jpeg', 'png'],
+        onErrorShape : function(message, element) {
+            alert(message+'=> width should be greater than height');  // \u{2260} is html symbol ≠	&#8800;	&#x2260; &ne; NOT EQUAL TO
+        },
+        onInvalidType : function(message, element) {
+            alert(message);
+        },
+        onSuccess : function(message, element) {
+            // alert(message);
+        }
+    });
+
+    $('#changeBlogLogoForm').submit( function(e) {
+        e.preventDefault();
+        var form = this;
+        $.ajax({
+            url         : $(form).attr('action'),
+            method      : $(form).attr('method'),
+            data        : new FormData(form),
+            processData : false,
+            dataType    : 'json',
+            contentType : false,
+            beforeSend  : function() {},
+            success     : function(data){
+                toastr.remove();
+                if (data.status == 1) {
+                    toastr.success(data.msg);
+                    $(form)[0].reset();
+                    Livewire.dispatch('updateTopHeader');
+                } else {
+                    toastr.error(data.msg);
+                }
+            }
+        });
+    } );
+    
+  </script>
+
+@endpush